refactor(docs): navbar dx
